Fix: Ensure system stubs are prioritized in Kotlin classpath
This addresses an issue I found while attempting to convert the build to use Android Studio / AndroidX dependencies. It uses a slightly newer Kotlin compiler and I suspect that something changed related to how Kotlin assembles it's classpath.

This commit modifies the `configureJavaKotlin.kt` script to ensure that system stubs are consistently prioritized in the classpath for both Kotlin and Java compilation tasks.

For KotlinCompile tasks, the system stubs are now added as the first library in the classpath.
For JavaCompile tasks, the system stubs are added as a task input with path normalization and then prepended to the task's classpath. This change addresses potential issues where other dependencies might overshadow the system stubs.

Test: ./build-and-test.sh
Relnote: Improve compatibility with Kotlin 2.1.20
Change-Id: Ib42d9f1fe41ecdc4f8f83aa6e7177518c66ce64b