Minor fixes: Android Plugin and gemini settings
Update Gemini CLI settings.json to new format (auto-formatted by Gemini itself)

Switch Android Gradle Plugin to be a compileOnly dependency of buildLogic
This is the correct way to use it as a dependency of a plugin.

Test: presubmit
RelNote: Corrected imethod for using the Android Gradle Plugin in a Gradle build plugin.
Change-Id: I28ffd26fb1a8fbfe63d5fe9ba9df4e08d4c2598f