Add the SHOW_MEDIA_PLAYBACK action to the Intent
For sources for which we `createMediaSessionIntent()` (those without
PendingIntent or non-MediaCenter apps), use the MEDIA_SHOW_PLAYBACK_VIEW
Intent action to indicate to the app that they should open to their now
playing view. It is on apps to react to this Intent Action. This is 1:1
with the Media card launch behavior in AAP and is necessary since CAL apps do
not have a mini player bar in their Browse view like MediaCenter apps
do.

Deprecate the `getIntent()` function in favor of `launchActivity(Context)`

Bug: 455949944
Test: build car-media-common, try on emulator with TestMediaCalApp
Relnote: "1. Add the SHOW_MEDIA_PLAYBACK action to MediaSession intents created
within MediaSource.
2. Deprecate the getIntent method in favor of
launchActivity"

Change-Id: I75e82404462e6a63248d968324d6635d720bf277